Use the browser default options when not in a Cordova environment
On the mobile devices we need to control more about how the inappbrowser should
look like, but that doesn't mean we should be mucking around with the regular
browser window outside of Cordova.

The behavior for the mobile app hasn't changed, but on desktop browsers we
will now honor the users preferences (open in window/tab, toolbars etc).